Backend/apps/push/serializers.py
from rest_framework import serializers

from .models import PushSubscription


class PushSubscriptionSerializer(serializers.ModelSerializer):
    class Meta:
        model = PushSubscription
        fields = ["id", "endpoint", "keys", "created_at"]
        read_only_fields = ["created_at"]
