Backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.6,<5.0
channels>=4.2,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
python-dotenv>=1.0,<2.0
Pillow>=11.0,<12.0
ruff>=0.9,<1.0
